+++ src/CMakeLists.txt
@@ -1,4 +1,4 @@
 # compile
 #add_subdirectory (test_prj)
-add_subdirectory (ns_solver)
-#add_subdirectory (upwind)
+#add_subdirectory (ns_solver)
+add_subdirectory (upwind)
